Reduced false positive matches for simple emote codes in SpamFilterService
The issue for a single emote being detected as multiple within the GetTotalDefaultEmoteCount
method still persists, and I currently can't think of a good solution.
Example: :laughing: will be detected as 2 matches. Some other emotes will be detected as
2 or even 4 matches too.
A possible solution would be to annotate the number of matches next to the emote value