Add logic to counteract multimatches for single emotes
The approach is to store the amount of matches each single emote
has if checked against a regex matches, iterateing through all emotes
as regex search strings. This value is then used as divisor if a match for that
element is found. It is not entirely accurate yet, but should be a more accurate
solution than before.